add an easy master creation procedure
